src/main/java/com/excilys/cdb/mapper/CompanyMapper.java
package main.java.com.excilys.cdb.mapper;

import java.sql.ResultSet;
import java.sql.SQLException;

import org.springframework.jdbc.core.RowMapper;

import main.java.com.excilys.cdb.model.Company;

/**
 * Company mapper.
 * @author Aurelien Denoize
 */
public class CompanyMapper implements RowMapper<Company> {

    @Override
    public Company mapRow(ResultSet resultSet, int i) throws SQLException {
        Company company = new Company();
        company.setId(resultSet.getLong("id"));
        company.setName(resultSet.getString("name"));

        return company;
    }

}
